requirements.txt
openai
python-dotenv
py-cord
wikipedia
langchain
google-search-results
wolframalpha
newsapi-python